fix: added quast speedup
fix: added no-sv, no-snps, and threading to quast to allow for speedup
feat: added cromwell config with memory scaling feature. However, we
cannot use cromwell as it only supports WDL version 1.0 or draft-2 which
would require some refactoring of main workflow code, so I suggest we
table this for now. Crowmwell is slowly adapting WDL version 1.1 and
plans to release it this year.